Allowed users to give themselves roles
